+++ CHANGELOG.md
@@ -12,7 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-* Nothing
+* ```run-pip-check.sh``` now runs in CircleCI pipeline (although it's a no-op
+  so functionality doesn't do anything but once the no-op is replaced we now
+  won't have to remember to uncomment the check in the pipeline)
 
 ### Removed